Started implementing mouse button events' (#9)
* Started implementing mouse button events'

* Event system almost finished, some tweaks are needed in the future

+++ src/main.odin
@@ -4,8 +4,10 @@ import "core:log"
 import "raytracer"
 
 main :: proc() {
-	context.logger = log.create_console_logger(opt = {.Level, .Terminal_Color})
-	defer log.destroy_console_logger(context.logger)
+	when ODIN_DEBUG {
+		context.logger = log.create_console_logger(opt = {.Level, .Terminal_Color})
+		defer log.destroy_console_logger(context.logger)
+	}
 	app := &raytracer.Application{}
 	err := raytracer.application_init(app, 1920, 1080, "Raytracer")
 	if err != nil {
